Added restart, both shutdown and restart only work on FOMOS/Linux.

+++ osShellScripts/shutdown.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
-echo "Shutting down..."
-shutdown -P now
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    echo "Shutting down..."
+    shutdown -P now
+else
+    echo "Shutdown doesn't work on this OS"
+fi
